Front-end routes that took forever to understand and create, sorry!
